Refactor: move combinations function to s2.py and update s3.py to use it
